+++ src/Core/Bind.hs
@@ -66,8 +66,7 @@ binder lv term ctx vars = case term of
   Pri p       -> Pri p
   Log s x     -> Log (binder lv s ctx vars) (binder lv x ctx vars)
   Met k t c   -> Met k (binder lv t ctx vars) (map (\x -> binder lv x ctx vars) c)
-  Pat s m c   -> undefined
-
+  Pat s m c   -> error "unreachable"
     -- -- Since Pat doesn't bind with HOAS, keep as Var
     -- let s'     = map (\x -> binder lv x ctx vars) s
         -- m'     = map (\(k,x) -> (k, binder lv x ctx vars)) m
